Split include.hpp
